fix autoloading for proof general and coq-mode
Goal:

- do not load proof-general/company-coq-mode on startup
- load proof-general/company-coq-mode when I open coq files
- execute the :config callback when the packages are loaded

Why is it so difficult? Because proof-general autoloads coq-mode and
doesn't load fully (provide 'proof-general) when only coq-mode is
autoloaded. Therefore, the use-package's use of eval-after-load
callback that listen for 'proof-general feature never run.

The fix is to move the :config block into a separate
with-eval-after-load callback and have it called whenever coq-mode is
loaded.