attempt to fix apheleia destroying elixir files
Apheleia will try to interrupt existing apheleia process when there is
one runinng. When Elixir's mix format was interrupted, it prints
something like the follows:

      BREAK: (a)bort (A)bort with dump (c)ontinue (p)roc info (i)nfo
             (l)oaded (v)ersion (k)ill (D)b-tables (d)istribution

The content is in turn copied to the current buffer and the file is
saved. If not careful now, the whole file would be lost.